move abstract setting screens to `base` package
