+++ src/main/java/meteordevelopment/meteorclient/gui/screens/settings/BlockListSettingScreen.java
@@ -6,6 +6,7 @@
 package meteordevelopment.meteorclient.gui.screens.settings;
 
 import meteordevelopment.meteorclient.gui.GuiTheme;
+import meteordevelopment.meteorclient.gui.screens.settings.base.CollectionListSettingScreen;
 import meteordevelopment.meteorclient.gui.widgets.WWidget;
 import meteordevelopment.meteorclient.mixin.IdentifierAccessor;
 import meteordevelopment.meteorclient.settings.BlockListSetting;
