+++ src/main/java/meteordevelopment/meteorclient/gui/screens/settings/StorageBlockListSettingScreen.java
@@ -7,6 +7,7 @@ package meteordevelopment.meteorclient.gui.screens.settings;
 
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
 import meteordevelopment.meteorclient.gui.GuiTheme;
+import meteordevelopment.meteorclient.gui.screens.settings.base.CollectionListSettingScreen;
 import meteordevelopment.meteorclient.gui.widgets.WWidget;
 import meteordevelopment.meteorclient.settings.Setting;
 import meteordevelopment.meteorclient.settings.StorageBlockListSetting;
