+++ src/main/java/meteordevelopment/meteorclient/gui/screens/settings/base/CollectionListSettingScreen.java
@@ -3,7 +3,7 @@
  * Copyright (c) Meteor Development.
  */
 
-package meteordevelopment.meteorclient.gui.screens.settings;
+package meteordevelopment.meteorclient.gui.screens.settings.base;
 
 import meteordevelopment.meteorclient.gui.GuiTheme;
 import meteordevelopment.meteorclient.gui.WindowScreen;
